# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ВСМ Тренажёр API",
    "description": "Backend геймифицированного тренажёра проводников высокоскоростной магистрали: движок сценариев, начисление очков компетенций, геймификация, аналитика и API для HR/LMS.\n\nАвторизация пользователей — Bearer JWT (`POST /api/v1/auth/login`), внешних систем — заголовок `X-API-Key`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Игровой профиль: уровень, баллы, достижения, челленджи",
        "operationId": "profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/reset": {
      "post": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Сбросить свою статистику: попытки, очки опыта, достижения, челленджи",
        "operationId": "reset_profile_api_v1_profile_reset_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/leaderboard": {
      "get": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Рейтинг по активным баллам (сгорают через 30 дней)",
        "operationId": "get_leaderboard_api_v1_leaderboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "brigade",
                "depot",
                "company"
              ],
              "type": "string",
              "default": "brigade",
              "title": "Scope"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Leaderboard Api V1 Leaderboard Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Уведомления (новые сценарии, челленджи, сгорающие баллы…)",
        "operationId": "notifications_api_v1_notifications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/read": {
      "post": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Отметить уведомления прочитанными",
        "operationId": "mark_read_api_v1_notifications_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkRead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/me": {
      "get": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Моя аналитика: компетенции, выводы, рекомендации",
        "operationId": "my_analytics_api_v1_analytics_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response My Analytics Api V1 Analytics Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/team": {
      "get": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Аналитика по бригадам и сотрудникам (тренер)",
        "operationId": "team_api_v1_analytics_team_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Team Api V1 Analytics Team Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/users/{user_id}": {
      "get": {
        "tags": [
          "Профиль и геймификация"
        ],
        "summary": "Аналитика конкретного сотрудника (тренер)",
        "operationId": "employee_analytics_api_v1_analytics_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Employee Analytics Api V1 Analytics Users  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AchievementOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "earned_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earned At"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "target": {
            "type": "integer",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "description",
          "icon",
          "earned_at",
          "progress",
          "target"
        ],
        "title": "AchievementOut"
      },
...
      "ChallengeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "ends": {
            "type": "string",
            "title": "Ends"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "target": {
            "type": "integer",
            "title": "Target"
          },
          "reward_xp": {
            "type": "integer",
            "title": "Reward Xp"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "ends",
          "progress",
          "target",
          "reward_xp",
          "completed_at"
        ],
        "title": "ChallengeOut"
      },
...
      "CompetencyOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "mastery": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mastery"
          },
          "samples": {
            "type": "integer",
            "title": "Samples"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "description",
          "mastery",
          "samples",
          "status"
        ],
        "title": "CompetencyOut"
      },
...
      "Expiring": {
        "properties": {
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "points",
          "expires_at"
        ],
        "title": "Expiring"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkRead": {
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          }
        },
        "type": "object",
        "title": "MarkRead"
      },
...
      "NotificationList": {
        "properties": {
          "unread": {
            "type": "integer",
            "title": "Unread"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "unread",
          "items"
        ],
        "title": "NotificationList"
      },
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "link",
          "created_at",
          "read_at"
        ],
        "title": "NotificationOut"
      },
...
      "Profile": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "level": {
            "additionalProperties": true,
            "type": "object",
            "title": "Level"
          },
          "total_xp": {
            "type": "integer",
            "title": "Total Xp"
          },
          "active_points": {
            "type": "integer",
            "title": "Active Points"
          },
          "expiring": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Expiring"
              },
              {
                "type": "null"
              }
            ]
          },
          "streak_days": {
            "type": "integer",
            "title": "Streak Days"
          },
          "finished_attempts": {
            "type": "integer",
            "title": "Finished Attempts"
          },
          "achievements": {
            "items": {
              "$ref": "#/components/schemas/AchievementOut"
            },
            "type": "array",
            "title": "Achievements"
          },
          "challenges": {
            "items": {
              "$ref": "#/components/schemas/ChallengeOut"
            },
            "type": "array",
            "title": "Challenges"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/CompetencyOut"
            },
            "type": "array",
            "title": "Competencies"
          }
        },
        "type": "object",
        "required": [
          "user",
          "level",
          "total_xp",
          "active_points",
          "expiring",
          "streak_days",
          "finished_attempts",
          "achievements",
          "challenges",
          "competencies"
        ],
        "title": "Profile"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "login": {
            "type": "string",
            "title": "Login"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "employee_code": {
            "type": "string",
            "title": "Employee Code"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "brigade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brigade"
          },
          "depot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depot"
          }
        },
        "type": "object",
        "required": [
          "id",
          "login",
          "display_name",
          "employee_code",
          "role",
          "brigade",
          "depot"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT из POST /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}